+++ changelog.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Breaking
+- module bitcoin-jsonrpc-client-autoconfigure: dependencies not included in anymore
+
 ### Changes
 - upgrade: update spring-boot from v3.1.0 to v3.1.1
 - upgrade: update cln-grpc-client-core from v23.5.1 to v23.5.2